Refactored use buttonInteractionFunction to external component for reuse by multiple components
